Added Inherit trait and made uses inherited
