+++ src/generate.rs
@@ -1,6 +1,7 @@
 //! Defines the code translations for the various macro components
 
 use crate::block::*;
+use crate::inherit::Inherit;
 use proc_macro2::TokenStream;
 use quote::quote;
 
@@ -43,32 +44,7 @@ impl Generate for Block {
 impl Generate for Describe {
     fn generate(mut self, parent: Option<&Describe>) -> TokenStream {
         if let Some(ref parent) = parent {
-            self.properties.inherit(parent);
-
-            if let Some(ref parent_before) = &parent.before {
-                // Prepend parent's `before` code sequence
-                let before = if let Some(self_before) = self.before {
-                    parent_before
-                        .0
-                        .iter()
-                        .chain(self_before.0.iter())
-                        .cloned()
-                        .collect()
-                } else {
-                    parent_before.0.clone()
-                };
-
-                self.before = Some(BasicBlock(before));
-            }
-
-            if let Some(ref parent_after) = &parent.after {
-                // Append parent's `after` code sequence
-                if let Some(ref mut self_after) = &mut self.after {
-                    self_after.0.extend(parent_after.0.clone());
-                } else {
-                    self.after = Some(parent_after.clone());
-                }
-            }
+            self.inherit(parent);
         }
 
         // Generate corresponding `use` statements
@@ -98,28 +74,9 @@ impl Generate for Describe {
 /// sequences
 impl Generate for Test {
     fn generate(mut self, parent: Option<&Describe>) -> TokenStream {
-        let content = if let Some(ref parent) = parent {
-            self.properties.inherit(parent);
-
-            // Prepend parent's `before` code sequence
-            let mut content = if let Some(ref parent_before) = &parent.before {
-                parent_before.0.clone()
-            } else {
-                Vec::new()
-            };
-            // Add test's unique code sequence
-            content.extend(self.content.0);
-            // Append parent's `after` code sequence
-            let parent_after = if let Some(ref parent_after) = &parent.after {
-                parent_after.0.clone()
-            } else {
-                Vec::new()
-            };
-            content.extend(parent_after);
-            content
-        } else {
-            self.content.0
-        };
+        if let Some(ref parent) = parent {
+            self.inherit(parent);
+        }
 
         let BlockProperties {
             attributes,
@@ -127,6 +84,7 @@ impl Generate for Test {
             ident,
             return_type,
         } = self.properties;
+        let content = self.content.0;
 
         // Generate the outer attributes and optional `async` token for this test
         let (attr_tokens, async_token) = if is_async {
@@ -156,30 +114,6 @@ impl Generate for Test {
                     #(#content)*
                 }
             )
-        }
-    }
-}
-
-impl BlockProperties {
-    /// Inherits the properties of the parent Describe block
-    fn inherit(&mut self, parent: &Describe) {
-        // Prepend parent's attributes
-        self.attributes = parent
-            .properties
-            .attributes
-            .iter()
-            .chain(self.attributes.iter())
-            .cloned()
-            .collect();
-
-        // If parent is async, so is self
-        if parent.properties.is_async {
-            self.is_async = true;
-        }
-
-        // If self doesn't have a return type, copy its parent's
-        if self.return_type.is_none() {
-            self.return_type = parent.properties.return_type.clone()
         }
     }
 }
